requirements.txt
langchain>=0.2.0
langgraph>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-requests
